requirements.txt
uvicorn
scikit-learn==1.6.1
joblib
transformers
sounddevice
numpy
faster-whisper
torch
torchvision
torchaudio
fastapi
sqlalchemy
passlib
uvcorn
psycopg2
pydantic[email]
pydantic
python-dotenv
langchain
langchain-google-genai